requirements.txt
# Python dependencies for Cash Points Telegram Bot
# This file is used for Railway deployment

# Telegram Bot API
python-telegram-bot==20.7

# Firebase Admin SDK
firebase-admin==6.4.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Date and time handling
python-dateutil==2.8.2

# JSON handling
jsonschema==4.20.0

# Logging and debugging
colorlog==6.8.0

# Async support
aiohttp==3.9.1

# Rate limiting
ratelimit==2.2.1

# Flask for Railway webhook
Flask==3.0.0

# Gunicorn for production server
gunicorn==21.2.0
